refactor: AbstractController 추가, 구현체에서 method 분기문 제거
