+++ tomcat/src/main/java/nextstep/jwp/controller/AbstractController.java
@@ -1,0 +1,27 @@
+package nextstep.jwp.controller;
+
+import nextstep.jwp.exception.UnsupportedMethodException;
+import org.apache.coyote.Controller;
+import org.apache.coyote.http11.message.request.HttpRequest;
+import org.apache.coyote.http11.message.request.requestline.Method;
+import org.apache.coyote.http11.message.response.HttpResponse;
+
+public abstract class AbstractController implements Controller {
+
+    @Override
+    public HttpResponse service(final HttpRequest httpRequest) throws Exception {
+        if (httpRequest.isMethod(Method.GET)) {
+            return doGet(httpRequest);
+        }
+
+        if (httpRequest.isMethod(Method.POST)) {
+            return doPost(httpRequest);
+        }
+
+        throw new UnsupportedMethodException();
+    }
+
+    protected abstract HttpResponse doGet(HttpRequest httpRequest) throws Exception;
+
+    protected abstract HttpResponse doPost(HttpRequest httpRequest) throws Exception;
+}
