+++ tomcat/src/main/java/nextstep/jwp/controller/ResourceController.java
@@ -1,14 +1,21 @@
 package nextstep.jwp.controller;
 
-import org.apache.coyote.Controller;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import nextstep.jwp.exception.UnsupportedMethodException;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.response.HttpResponse;
 
-public class ResourceController implements Controller {
+public class ResourceController extends AbstractController {
 
     @Override
-    public HttpResponse service(final HttpRequest httpRequest) throws Exception {
+    protected HttpResponse doGet(final HttpRequest httpRequest) throws IOException, URISyntaxException {
         return HttpResponse.ofResource(httpRequest.getPath());
+    }
+
+    @Override
+    protected HttpResponse doPost(final HttpRequest request) {
+        throw new UnsupportedMethodException();
     }
 
     @Override
